TimeZoneFinder

Documentation for TimeZoneFinder.

API

TimeZoneFinder.timezone_at — Function

timezone_at(latitude, longitude)

Get the timezone at the given latitude and longitude.

julia> timezone_at(52.5061, 13.358)
Europe/Berlin (UTC+1/UTC+2)

Note

The library always uses the same version of tzdata currently used by TimeZones.

Returns a TimeZone instance if latitude and longitude correspond to a known timezone, otherwise nothing is returned.

Implementation details

Artifact

This module is based upon OpenStreetMap data, which has been compiled into shape files by timezone-boundary-builder. We define an Artifact for each release (since 2021c) of these raw JSON shape files.

When a new upstream release is made, a new artifact will be created by a package maintainer. Refer to the artifact_build directory for a helper script.

Parsed cache

This raw data is provided in JSON format, so the first time a package uses it, it is parsed (which can take tens of seconds). Subsequently, a serialized binary version in a scratch space is loaded, which is much faster. This cache is re-used so long as the package and Julia versions remain the same. After an upgrade the cache will be re-generated, which can cause a one-off latency of a few tens of seconds.

Parsed format

The parsed data contains multiple (about 1000) polygons, each with a correpsonding time-zone. Each polygon is stored as a PolyArea along with its axis-aligned bounding box. This bounding box is used to speed up checks for containment.

Lookup

Given a (latitude, longitude) point, we perform a linear scan over all polygons. For each polygon we check for containment, using the bounding box to quickly dismiss polygons that are far away.

Note

In the future, performance might be improved further by building a more fine-grained spatial index.