TimeZoneFinder

Documentation for TimeZoneFinder.

API

TimeZoneFinder._cache_path — Method

Serialized data file path for this version.

TimeZoneFinder._scratch_dir — Method

_scratch_dir(release)

Get the scratch directory path in which the serialized mapping data will be kept.

TimeZoneFinder.generate_data — Method

Generate the timezone map data from the artifact identified by release.

TimeZoneFinder.load_data — Method

load_data(release)

Load timezone map data for release.

This is memoized, such that the data is only read from disk once within the lifetime of the Julia process.

TimeZoneFinder.timezone_at — Method

timezone_at(latitude, longitude; release="2021c")

Get the timezone at the given latitude and longitude.

julia> timezone_at(52.5061, 13.358)
Europe/Berlin (UTC+1/UTC+2)

Returns a TimeZone instance if latitude and longitude correspond to a known timezone, otherwise nothing is returned.

Implementation details

This module is based upon OpenStreetMap data, which has been compiled into shape files by timezone-boundary-builder. We define an Artifact for each release (since 2021c) of these shape files.

This raw data is provided in JSON format, so the first time a package uses it, it is parsed (which can take tens of seconds). Subsequently, a serialized binary version in a scratch space is loaded, which is much faster. This cache is re-used so long as the package and Julia versions remain the same. After an upgrade the cache will be re-generated, which can cause a one-off latency of a few tens of seconds.

Caveats

The current implementation aims to be simple, however there is scope for further optimisation.

Finding a timezone currently involves a linear scan over a list of about 1000 polygons. For each polygon, which may have tens of thousands of line segments, we perform a containment check for the point. This could be made significantly more performant by adding an appropriate spatial index over the polygons, which would reduce the number of polygons that have to be checked.